I did not atomic commit
for this dashboard site ive decided to commondeer this project into a portfolio. I went with a purple
color scheme and did the rough outline of what the project wanted. I used some stock svgs and haven't
filled much of the written content out yet. its mostly a static unreactive site but i think it will
be good to have something to put in applications.